+++ tileserver/maptiles/mbtilesgenerator_sqlite3.go
@@ -30,16 +30,8 @@ func NewTileDbSqlite(path string) *TileDbSqlite3 {
 	queries := []string{
 		"PRAGMA journal_mode = OFF",
 		"CREATE TABLE IF NOT EXISTS layers(layer_name TEXT PRIMARY KEY NOT NULL)",
-		"CREATE TABLE IF NOT EXISTS metadata (name TEXT PRIMARY KEY NOT NULL, value TEXT NOT NULL, layer_name TEXT NOT NULL)",
+		"CREATE TABLE IF NOT EXISTS metadata (name TEXT NOT NULL, value TEXT NOT NULL, layer_name TEXT NOT NULL)",
 		"CREATE TABLE IF NOT EXISTS tiles (layer_id INTEGER, zoom_level INTEGER, tile_column INTEGER, tile_row INTEGER, tile_data blob, PRIMARY KEY (layer_id, zoom_level, tile_column, tile_row))",
-		// "REPLACE INTO metadata VALUES('name', 'go-mapnik cache file')",
-		// "REPLACE INTO metadata VALUES('type', 'overlay')", //baselayer
-		// "REPLACE INTO metadata VALUES('version', '1')",
-		// "REPLACE INTO metadata VALUES('description', 'Compatible with MBTiles spec 1.2. However, this file may contain multiple overlay layers, but only the layer called default is exported as MBtiles')",
-		// "REPLACE INTO metadata VALUES('format', 'png')",
-		// "REPLACE INTO metadata VALUES('bounds', '-180.0,-85,180,85')",
-		// "REPLACE INTO metadata VALUES('attribution', 'sjsafranek')",
-		// "INSERT OR IGNORE INTO layers(layer_name) VALUES('default')",
 	}
 
 	for _, query := range queries {
@@ -185,9 +177,51 @@ func (self *TileDbSqlite3) fetch(r TileFetchRequest) {
 	r.OutChan <- result
 }
 
+// AddLayerMetadata adds metadata t0 metadata table
+func (self *TileDbSqlite3) AddLayerMetadata(lyr string, stylesheet string) {
+
+	check_query := "SELECT EXISTS(SELECT * FROM metadata WHERE name='name' AND layer_name='" + lyr + "')"
+
+	insert_queries := []string{
+		"INSERT INTO metadata(name, value, layer_name) VALUES('name', '" + lyr + "', '" + lyr + "')",
+		"INSERT INTO metadata(name, value, layer_name) VALUES('source', '" + stylesheet + "', '" + lyr + "')",
+		"INSERT INTO metadata(name, value, layer_name) VALUES('type', 'overlay', '" + lyr + "')",
+		"INSERT INTO metadata(name,value,layer_name) VALUES('version', '1', '" + lyr + "')",
+		"INSERT INTO metadata(name,value,layer_name) VALUES('description', 'Compatible with MBTiles spec 1.2.', '" + lyr + "')",
+		"INSERT INTO metadata(name,value,layer_name) VALUES('format', 'png', '" + lyr + "')",
+		"INSERT INTO metadata(name,value,layer_name) VALUES('bounds', '-180.0,-85,180,85', '" + lyr + "')",
+		"INSERT INTO metadata(name,value,layer_name) VALUES('attribution', 'sjsafranek', '" + lyr + "')",
+	}
+
+	if !self.rowExists(check_query) {
+		Ligneous.Info("Adding metadata for ", lyr)
+		for _, query := range insert_queries {
+			_, err := self.db.Exec(query)
+			if err != nil {
+				Ligneous.Error("Error adding metadata to db", err.Error())
+				Ligneous.Debug(query, "\n")
+			}
+		}
+	}
+
+}
+
+// rowExists checks if row exists in table
+func (self *TileDbSqlite3) rowExists(query string) bool {
+	var exists bool
+	err := self.db.QueryRow(query).Scan(&exists)
+	if nil != err {
+		Ligneous.Error(err)
+	}
+	return exists
+}
+
 // MetaDataHandler gets metadata from database.
 func (self *TileDbSqlite3) MetaDataHandler(lyr string) map[string]string {
-	rows, _ := self.db.Query("SELECT * FROM metadata WHERE layer_name=?", lyr)
+	rows, err := self.db.Query("SELECT name, value FROM metadata WHERE layer_name=?", lyr)
+	if nil != err {
+		Ligneous.Error(err)
+	}
 	metadata := make(map[string]string)
 	for rows.Next() {
 		var name string